performer and client side progress: 70%

+++ application/views/performer/package.php
@@ -1,31 +1,40 @@
 <div class="container">
 
-<p class="h3 py-3 yellow-brown text-center">Your Packages</p>
+<?php echo $this->session->flashdata('success_message') ? $this->message_model->success_message() : '' ?>
+<?php echo $this->session->flashdata('danger_message') ? $this->message_model->danger_message() : '' ?>
 
-<div class="card-columns">
-	<?php foreach($packages as $p): ?>
-	<div class="card">
-		<div class="card-header bg-primary">
-			<p class="h5 text-center text-white"><?php echo $p['package_name'] ?></p>
+
+<?php if($packages): ?>
+	<p class="h3 py-3 yellow-brown text-center">Your Packages</p>
+	<div class="card-columns">
+		<?php foreach($packages as $p): ?>
+		<div class="card">
+			<div class="card-header bg-primary">
+				<p class="h5 text-center text-white"><?php echo $p['package_name'] ?></p>
+			</div>
+			<ul class="list-group list-group-flush">
+				<li class="list-group-item"><?php echo $p['details'] ?></li>
+				<li class="list-group-item">₱ <?php echo $p['price'] ?></li>
+				<li class="list-group-item">
+					<div class="row">
+						<div class="col-md-8 ">
+							<small>Created: <?php echo date('h:i:s a M d, y', strtotime($p['created_at'])) ?></small>
+							<small>Updated: <?php echo date('h:i:s a M d, y', strtotime($p['updated_at'])) ?></small>
+						</div>
+						<div class="col-md-4 ">
+							<a href="p_package_edit_page/<?php echo $p['package_id'] ?>" class="btn btn-primary text-white"><i class="fas fa-pen"></i></a>
+							<a href="p_package_delete/<?php echo $p['package_id'] ?>" class="btn btn-danger text-white"><i class="fas fa-trash"></i></a>
+						</div>
+					</div>
+				</li>
+			</ul>
 		</div>
-		<ul class="list-group list-group-flush">
-			<li class="list-group-item"><?php echo $p['details'] ?></li>
-			<li class="list-group-item">₱ <?php echo $p['price'] ?></li>
-			<li class="list-group-item">
-				<div class="row">
-					<div class="col-md-8 ">
-						<small>Created: <?php echo date('H:i:s a M d, y', strtotime($p['created_at'])) ?></small>
-						<small>Updated: <?php echo date('H:i:s a M d, y', strtotime($p['updated_at'])) ?></small>
-					</div>
-					<div class="col-md-4 ">
-						<a href="p_package_edit_page/<?php echo $p['package_id'] ?>" class="btn btn-primary text-white"><i class="fas fa-pen"></i></a>
-						<a href="p_package_delete/<?php echo $p['package_id'] ?>" class="btn btn-danger text-white"><i class="fas fa-trash"></i></a>
-					</div>
-				</div>
-			</li>
-		</ul>
+		<?php endforeach; ?>
 	</div>
-	<?php endforeach; ?>
-</div>
-	
+<?php else: ?>
+	<div class="alert alert-secondary my-3">
+		<p class="h3 text-center">No Package/s Found</p>
+	</div>
+<?php endif; ?>
+
 </div>
